requirements.txt
tabula-py >= 2.3.0
PyPDF4 == 1.27.0
PyPDF2 == 